Add kOS "data center terminal" processors to all launch pads.
The idea is to have a processor with unlimited storage (read: runs directly from the archive)
but is tied to the ground, to handle (possibly complex) pre-launch tasks before
handing off to the weaker on-board processor.

Example: Data center automatically calculates an optimal gravity turn on launch,
and then hands off the initial pitch to the onboard processor which simply
executes a pitch program to the initial pitch and follows prograde.